third_party/WebKit/LayoutTests/editing/execCommand/format-block-with-braces.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable>|Format Me</div>',
        'formatBlock <h1>',
        '<div contenteditable><h1>|Format Me</h1></div>'),
        'foramtBlcok accept tag name with html brackets');
</script>
